Dependencies removed from results aggregator, baseline in GP foraging path, apptainer files added

+++ charts/aggregator.py
@@ -3,51 +3,85 @@ import os
 import sys
 sys.path.insert(0, '..')
 
-from params import eaParams
 from pathlib import Path
 
 class Aggregator():
 
     def __init__(self):
-        self.params = eaParams()
+
+        self.objectives = ["density", "nest", "food", "idensity", "inest", "ifood", "foraging"]
+
+        with open("../path.txt", "r") as f:
+            for line in f:
+                data = line.split(":")
+                if data[0] == "home": self.home_path = data[1][0:-1]
+                if data[0] == "local": self.local_path = data[1][0:-1]
+                if data[0] == "shared": self.shared_path = data[1][0:-1]
+
         self.algorithm = "gp"
-        self.objective = self.params.objectives[0]
+        self.using_repertoire = False
+        self.repertoire_type = "qd"
+        self.repertoire_size = 1
+        self.objective = self.objectives[0]
         self.runs = 0
         self.generations = 0
+
+        self.cancelled = False
+
         self.configure()
-        self.cancelled = "repro" in self.params.shared_path
+        self.cancelled = self.cancelled or "repro" in self.shared_path
 
     def configure(self):
-        with open(self.params.local_path+"/config.txt", 'r') as f:
+        permitted = ["algorithm", "using_repertoire", "repertoire_type", "repertoire_size", "objective", "runs", "generations"]
+        with open(self.local_path+"/config.txt", 'r') as f:
             for line in f:
                 data = line.split()
                 if len(data) > 0:
-                    print(line[0:-1])
-                    self.update(data)
+                    if data[0] in permitted:
+                        print(line[0:-1])
+                        self.update(data)
+                    else:
+                        print("\nconfig entry not recognised: "+data[0]+"\n")
+                        self.cancelled = True
 
     def update(self, data):
-        if data[0] == "algorithm": self.algorithm = data[1]
-        if data[0] == "objective": self.objective = self.params.objectives[int(data[1])]
+        if data[0] == "algorithm":
+            algorithms = ["gp", "qdpy", "cma-es"]
+            if data[1] in algorithms:
+                self.algorithm = data[1]
+            else:
+                print("\nalgorithm not recognised: "+data[1]+"\n")
+                self.cancelled = True
+        if data[0] == "using_repertoire": self.using_repertoire = True if data[1] == "True" else False
+        if data[0] == "repertoire_type": self.repertoire_type = data[1]
+        if data[0] == "repertoire_size": self.repertoire_size = int(data[1])
+        if data[0] == "objective": self.objective = self.objectives[int(data[1])]
         if data[0] == "runs": self.runs = int(data[1])
         if data[0] == "generations": self.generations = int(data[1])
 
+    def addRepertoireDir(self, path):
+        if self.objective == "foraging" and self.algorithm == "gp":
+            if self.using_repertoire:
+                path += "/"+self.repertoire_type+str(self.repertoire_size)
+            else:
+                path += "/baseline"
+        return path
+
     def getResults(self):
 
-        print("\nreading...")
+        path = self.shared_path+"/"+self.algorithm+"/"+self.objective
+        path = self.addRepertoireDir(path)
+        print("\nreading from "+path+"\n")
+
         results_per_query = []
 
         for query in self.queries:
-
-            path = self.params.shared_path+"/"+self.algorithm+"/"+self.objective+"/"
-            if self.objective == "foraging":
-                path += self.params.repertoire_type+str(self.params.repertoire_size)+"/"
-            path += query
 
             entries = []
 
             for i in range(1, self.runs + 1):
 
-                filename = path+str(self.generations)+"-"+str(i)+".csv"
+                filename = path+"/"+query+str(self.generations)+"-"+str(i)+".csv"
 
                 if os.path.exists(filename):
 
@@ -76,8 +110,15 @@ class Aggregator():
 
     def writeCSV(self, results_per_query):
 
-        print("\nwriting...")
-        Path(self.params.home_path+"/"+self.algorithm+"/").mkdir(parents=False, exist_ok=True)
+        Path(self.home_path+"/"+self.algorithm+"/").mkdir(parents=False, exist_ok=True)
+        
+        filepath = self.home_path+"/"+self.algorithm+"/"+self.objective
+        Path(filepath).mkdir(parents=False, exist_ok=True)
+
+        filepath = self.addRepertoireDir(filepath)
+        Path(filepath).mkdir(parents=False, exist_ok=True)
+
+        print()
         confirmed = False
 
         for i in range(len(self.queries)):
@@ -87,19 +128,19 @@ class Aggregator():
 
             if len(entries) > 0:
 
-                filename = self.params.home_path+"/"+self.algorithm+"/"+query+str(self.generations)+"-"+self.objective
-                if self.objective == "foraging":
-                    filename += "-"+self.params.repertoire_type+str(self.params.repertoire_size)
-                filename += ".csv"
+                filename = filepath+"/"+query+str(self.generations)+".csv"
 
                 if not confirmed and os.path.exists(filename):
-                    test = input("Output files already exist at "+filename+"\nContinue? (y/N)\n")
+
+                    test = input("Output files already exist at "+filepath+"\nContinue? (y/N)\n")
                     if test == "y":
                         confirmed = True
                         print()
                     else:
+                        print("\ncancelled\n")
                         return False
 
+                print("Writing "+filename)
                 with open(filename, "w") as f:
                     for entry in entries:
                         f.write(entry)
@@ -108,18 +149,21 @@ class Aggregator():
 
     def removeOldFiles(self):
 
-        print("\ndeleting...")
+        print()
+        test = input("Delete input files? (y/N)\n")
+        if test != "y":
+            print("\ncancelled\n")
+            return
+
+        path = self.shared_path+"/"+self.algorithm+"/"+self.objective
+        path = self.addRepertoireDir(path)
+        print("\ndeleting from "+path)
+
         for query in self.queries:
-
-            path = self.params.shared_path+"/"+self.algorithm+"/"+self.objective+"/"
-            if self.objective == "foraging":
-                path += self.params.repertoire_type+str(self.params.repertoire_size)+"/"
-            path += query
-
 
             for i in range(1, self.runs + 1):
 
-                filename = path+str(self.generations)+"-"+str(i)+".csv"
+                filename = path+"/"+query+str(self.generations)+"-"+str(i)+".csv"
 
                 if os.path.exists(filename):
                     os.remove(filename)
@@ -145,4 +189,4 @@ if __name__ == "__main__":
         results_saved = aggregator.writeCSV(results_per_query)
 
         if results_saved:
-            aggregator.removeOldFiles()
+           aggregator.removeOldFiles()
